Move CholmodSimplicialLDLT2 out of AstrometryFit and cleanup #includes
Start switching PhotometryFit to use Astrometry's Cholmod code
